requirements.txt
Django==4.2
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
django-admin-volt
django_compressor
django-libsass
django-froala-editor
djangorestframework-recursive